# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ZenFM API
  version: 0.1.0
  description: |
    The local, single-owner ZenFM HTTP API. Browser requests authenticate with
    an opaque `zenfm_session` cookie and send the session-bound value returned
    as `csrfToken` in `X-ZenFM-CSRF` for every mutation. Personal API tokens use
    `Authorization: Bearer`; their restricted scope excludes owner security,
    token management, update, and lifecycle operations.
servers:
  - url: /
paths:
...
  /api/v1/files/preview:
    get:
      operationId: previewFile
      description: Returns a bounded, safe preview representation for a regular file.
      security:
        - sessionCookie: []
        - personalToken: []
      parameters:
        - $ref: '#/components/parameters/Path'
        - in: query
          name: width
          schema:
            type: integer
            minimum: 16
            maximum: 4096
        - in: query
          name: height
          schema:
            type: integer
            minimum: 16
            maximum: 4096
      responses:
        '200':
          description: Preview stream.
          content:
            image/png:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
            text/html:
              schema:
                type: string
              description: Sanitized HTML only; scripts, event handlers, and external resources are removed.
            text/csv:
              schema:
                type: string
            text/markdown:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
            audio/*:
              schema:
                type: string
                format: binary
            video/*:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          $ref: '#/components/responses/TooLarge'
        '415':
          $ref: '#/components/responses/Unsupported'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zenfm_session
    personalToken:
      type: http
      scheme: bearer
      bearerFormat: opaque ZenFM personal token
  parameters:
...
    Path:
      in: query
      name: path
      required: true
      schema:
        $ref: '#/components/schemas/Path'
...
  responses:
...
    Unauthorized:
      description: Missing, expired, or revoked credentials.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Forbidden:
      description: Authenticated principal lacks access, setup is incomplete, CSRF failed, or origin was rejected.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooLarge:
      description: A configured body, upload, preview, or archive bound was exceeded.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unsupported:
      description: The target cannot be safely previewed or opened as regular content.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Path:
      type: string
      minLength: 1
      maxLength: 4096
      pattern: '^/'
      examples: [/Books]
...
    Problem:
      type: object
      additionalProperties: true
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference